Extract common members into a common base
Members related to file version and datamodels contained in files are
the same for all readers and some commonly available member functions
are also shared across all readers.
