+++ include/podio/SIOReader.h
@@ -3,8 +3,7 @@
 
 #include "podio/SIOBlock.h"
 #include "podio/SIOFrameData.h"
-#include "podio/podioVersion.h"
-#include "podio/utilities/DatamodelRegistryIOHelpers.h"
+#include "podio/utilities/ReaderCommon.h"
 
 #include <sio/definitions.h>
 
@@ -22,7 +21,7 @@ class CollectionIDTable;
 ///
 /// The SIOReader provides the data as SIOFrameData from which a podio::Frame
 /// can be constructed. It can be used to read files written by the SIOWriter.
-class SIOReader {
+class SIOReader : public ReaderCommon {
 
 public:
   /// Create an SIOReader
@@ -83,45 +82,10 @@ public:
   /// @param filename The path to the file to read from
   void openFile(const std::string& filename);
 
-  /// Get the build version of podio that has been used to write the current
-  /// file
-  ///
-  /// @returns The podio build version
-  podio::version::Version currentFileVersion() const {
-    return m_fileVersion;
-  }
-
-  /// Get the (build) version of a datamodel that has been used to write the
-  /// current file
-  ///
-  /// @param name The name of the datamodel
-  ///
-  /// @returns The (build) version of the datamodel if available or an empty
-  ///          optional
-  std::optional<podio::version::Version> currentFileVersion(std::string_view name) const {
-    return m_datamodelHolder.getDatamodelVersion(name);
-  }
-
   /// Get the names of all the available Frame categories in the current file.
   ///
   /// @returns The names of the available categores from the file
   std::vector<std::string_view> getAvailableCategories() const;
-
-  /// Get the datamodel definition for the given name
-  ///
-  /// @param name The name of the datamodel
-  ///
-  /// @returns The high level definition of the datamodel in JSON format
-  const std::string_view getDatamodelDefinition(std::string_view name) const {
-    return m_datamodelHolder.getDatamodelDefinition(name);
-  }
-
-  /// Get all names of the datamodels that are available from this reader
-  ///
-  /// @returns The names of the datamodels
-  std::vector<std::string> getAvailableDatamodels() const {
-    return m_datamodelHolder.getAvailableDatamodels();
-  }
 
 private:
   void readPodioHeader();
@@ -138,10 +102,6 @@ private:
 
   /// Table of content record where starting points of named entries can be read from
   SIOFileTOCRecord m_tocRecord{};
-  /// The podio version that has been used to write the file
-  podio::version::Version m_fileVersion{0};
-
-  DatamodelDefinitionHolder m_datamodelHolder{};
 };
 
 } // namespace podio
